New route to add projects. Project schema: Changed ID to int type for easier data read/write, may change to UUID in future.

+++ src/main/java/com/tkato/myKanBan/repository/ProjectRepository.java
@@ -1,8 +1,8 @@
 package com.tkato.myKanBan.repository;
 
 import org.springframework.data.repository.CrudRepository;
-import com.tkato.myKanBan.model.*;
+import com.tkato.myKanBan.model.Project;
 
-public interface ProjectRepository extends CrudRepository<Project, Long> {
+public interface ProjectRepository extends CrudRepository<Project, Integer> {
 
 }
